src/backend/requirements.txt
uvicorn==0.25.0
fastapi==0.108.0
python-multipart==0.0.9
fastapi-utils==0.7.0
openai==1.56.0
pydantic
python-dotenv==1.0.1
pyyaml==6.0.1
easydict==1.13
google-cloud-firestore